requirements.txt
django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-tenants
djangorestframework-simplejwt
PyJWT==2.8.0
drf-spectacular==0.27.1